requirements.txt
Flask
Flask-Session
python-dotenv
requests
waitress
openai
mysql-connector-python
psutil

